updated code with tiling, train and predicting walls, + some addings on extracting dem data

+++ 1_preprocessing/0_assemble_dtm_tile.py
@@ -40,6 +40,7 @@ import numpy as np
 import shapely
 from osgeo import ogr, gdal
 
+#%%
 km10training = gpd.read_file(r'V:\2022-03-31_Stendiger_EZRA\stonewalls_slks\data\pilot_training_data\training_10km_grid.gpkg')
 km1 = gpd.read_file(r'V:\2022-03-31_Stendiger_EZRA\stonewalls_slks\data\grids\dki_1km.gpkg')
 # km10 = gpd.read_file(r'V:\2022-03-31_Stendiger_EZRA\stonewalls_slks\data\grids\dki_10km.gpkg')
@@ -48,10 +49,22 @@ km1 = gpd.read_file(r'V:\2022-03-31_Stendiger_EZRA\stonewalls_slks\data\grids\dk
 km10training['geometry'] = km10training.apply(lambda x: x['geometry'].buffer(1), axis=1)
 
 tile_list = gpd.overlay(km1, km10training, how='intersection')['dki_1km'].tolist()
+
+#%%
+
+missing_list = r"V:\2022-03-31_Stendiger_EZRA\code\km1_missing_tiles.txt"
+
+tile_list = []
+
+### adapted this to get the tilis that were missing --> adapt to the data you want to extract from
+with open(missing_list, "r") as file:
+    for tile in file:
+        tile_list.append(tile.strip())
+
 print(len(tile_list), 'tiles....')
 
 #%%
-len(gpd.overlay(km1, km10training, how='intersection')['dki_1km'].unique().tolist())
+# len(gpd.overlay(km1, km10training, how='intersection')['dki_1km'].unique().tolist())
 
 #%%
 import shutil
@@ -71,13 +84,18 @@ for tile in tile_list:
 
     dtm_dest_tif = os.path.join(DTM_destination_dir, 'DTM_'+ tile +'.tif' )
     dsm_dest_tif = os.path.join(DSM_destination_dir, 'DSM_'+ tile +'.tif' )
-
-    if not os.path.exists(dtm_dest_tif):
-        shutil.copy(dtm_source_tif, dtm_dest_tif)
-        print('DTM file didnt exist: ', dtm_source_tif)
-    if not os.path.isfile(dsm_dest_tif):
-        shutil.copy(dsm_source_tif, dsm_dest_tif)
-        print('DSM file didnt exist: ', dsm_source_tif)
+    
+    if not os.path.exists(dtm_source_tif):
+        print("tile does not exist: ", tile)
+    else:
+        if not os.path.exists(dtm_dest_tif):
+            shutil.copy(dtm_source_tif, dtm_dest_tif)
+            print('DTM file didnt exist: ', dtm_source_tif)
+    if not os.path.exists(dsm_source_tif):
+        print("tile does not exist: ", tile)
+    else:
+        if not os.path.isfile(dsm_dest_tif):
+            shutil.copy(dsm_source_tif, dsm_dest_tif)
+            print('DSM file didnt exist: ', dsm_source_tif)
     # print('moving: ',tile, '...')
 print('finished')
-# %%
